requirements.txt
boto3==1.5.6
botocore==1.8.20
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
docutils==0.14
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
Mastodon.py==1.2.1
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
rq==0.9.2
s3transfer==0.1.12
six==1.11.0
urllib3==1.22
Werkzeug==0.13
